add instructions for viewing the map
Hey, this is amazing. Great work.  I got it running with my google id, but then I didn't know how to view it. I happened to click that reddit link, even though I'm on a mac, and I found the info there. I thought it would be good to have it in the read me. :D